speech_recognition/requirements.txt
aiofiles
argparse
fastapi
passlib[bcrypt]
python-jose[cryptography]
sqlalchemy
uvicorn[standard]